Functions/requirements.txt
azure-core==1.26.4
azure-cosmos==4.3.1
azure-functions==1.14.0
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
requests==2.30.0
six==1.16.0
typing_extensions==4.5.0
urllib3==2.0.2
